requirements.txt
# requests
grpcio
chirpstack-api